Remove typo file
