+++ app/backend/src/main/kotlin/com/gamelounge/backend/util/ConverterDTO.kt
@@ -148,8 +148,9 @@ object ConverterDTO {
         )
     }
 
-    fun convertBulkToUserGameRatingDTO(userGameRatings: List<UserGameRating>): List<UserGameRatingDTO>{
+    fun convertBulkToUserGameRatingDTO(userGameRatings: List<UserGameRating>): List<UserGameRatingDTO> {
         return userGameRatings.map { userGameRating -> converToUserGameRatingDTO(userGameRating) }
+    }
 
     fun convertBulkToEditedGameDTO(editedGames: List<RequestedEditingGame>): List<EditedGameDTO>{
         return editedGames.map { editedGame -> convertToEditedGameDTO(editedGame) }
